requirements.txt
chromadb==0.6.3
kneed==0.8.5
matplotlib==3.10.3
nltk==3.9.1
numpy==2.3.1
openai==1.93.0
pandas==2.3.0
protobuf==6.31.1
python-dotenv==1.1.1
scikit_learn==1.7.0
sentence_transformers==4.1.0
spacy==3.8.7
together==1.5.17
tqdm==4.67.1
transformers==4.52.3
umap==0.1.1
fastapi
